% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.jmb_analysis}
\alias{is.jmb_analysis}
\title{Is a JAGS Analysis}
\usage{
is.jmb_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'jmb_analysis'
}
