% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.jmb_code}
\alias{is.jmb_code}
\title{Is JAGS Code}
\usage{
is.jmb_code(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'jmb_code'
}

