% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmbr-package.R
\docType{package}
\name{jmbr-package}
\alias{jmbr}
\alias{jmbr-package}
\title{jmbr: Analyses Using JAGS}
\description{
Facilitates analyses using 'Just Another Gibbs Sampler'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/jmbr}
  \item Report bugs at \url{https://github.com/poissonconsulting/jmbr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
