% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Map}
\usage{
map(inits)
}
\arguments{
\item{inits}{A named list of initial values.}
}
\description{
Constructs a list identifying which parameters to fix (based on missing values in inits).
}
\seealso{
\code{\link[TMB]{MakeADFun}}
}

