% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tmb_analysis}
\alias{predict.tmb_analysis}
\title{Predict}
\usage{
\method{predict}{tmb_analysis}(object, new_data = data_set(object),
  new_expr = NULL, new_values = list(), term = "prediction",
  conf_int = FALSE, conf_level = 0.95, modify_new_data = NULL,
  parallel = getOption("mb.parallel", FALSE), quick = getOption("mb.quick",
  FALSE), quiet = getOption("mb.quiet", TRUE), beep = getOption("mb.beep",
  FALSE), ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{new_expr}{A string of R code specifying the predictive relationship.}

\item{new_values}{A named list of new or replacement values to pass to new_expr.}

\item{term}{A string of the term in new_expr.}

\item{conf_int}{A flag indicating whether to calculate confidence intervals.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{modify_new_data}{A single argument function to modify new data (in list form) immediately prior to calculating new_expr.}

\item{parallel}{A flag indicating whether to do predictions using parallel backend provided by foreach.}

\item{quick}{A flag indicating whether to quickly get unreliable values.}

\item{quiet}{A flag indicating whether to disable tracing information.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{...}{Additional arguments.}
}
\value{
The new data with the predictions.
}
\description{
Calculate predictions.
}
\details{
If \code{conf_int = TRUE} then the confidence intervals are calculated by profiling.
The profiling only takes uncertainty in the fixed parameters into account.
When profiling \code{new_expr} must be linear (the sum of a series of terms)
and can only include one function that encompasses the whole expression,
i.e., \code{prediction <- exp(bIntercept + bSlope * y)}.
}

