% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{remap_coef}
\alias{remap_coef}
\title{Remap Coef}
\usage{
remap_coef(coef, map)
}
\arguments{
\item{coef}{A coef table to add as 0 values that were fixed.}

\item{map}{A named list indicating the fixed values.}
}
\description{
Adds fixed values to coef table.
}

