% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{remap_estimates}
\alias{remap_estimates}
\title{Remap Estimates}
\usage{
remap_estimates(estimates, map)
}
\arguments{
\item{estimates}{A named list of the estimates to add as 0 values that were fixed.}

\item{map}{A named list indicating the fixed values.}
}
\description{
Adds fixed values to estimates.
}

