% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{remap_vector}
\alias{remap_vector}
\title{Remap Vector}
\usage{
remap_vector(x, y)
}
\arguments{
\item{x}{The vector of estimates to remap (added fixed values as 0)}

\item{y}{The vector indicating missing values.}
}
\description{
Remaps vector x based on missing values in y.
}

