% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmbr-package.R
\docType{package}
\name{jmbr-package}
\alias{jmbr}
\alias{jmbr-package}
\title{jmbr: Analyses Using JAGS}
\description{
Facilitates analyses using 'Just Another Gibbs Sampler'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/jmbr}
  \item Report bugs at \url{https://github.com/poissonconsulting/jmbr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
